Move to fine-grained locking on Element
This swaps around ownership somewhat, as we need to lock the element
while we use it.  We rely on atomic operations and eventual consistency
via eviction to maintain the lists in their correct form, trying to
avoid holding locks for multiple elements in the same thread.

Only having a thread own a single element lock, and having the eviction
lock be also run on a separate thread to requests, means we should be
safe from deadlock.

We never take the eviction lock when holding an element lock, and we
never take a second element lock when holding an element lock.  It is
safe to take an element lock while holding the eviction lock.